ajout protection route controller

+++ backend/controllers/Sauce.js
@@ -35,25 +35,38 @@ exports.getOneSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-  const sauceObject = req.file ?
-    {
-      ...JSON.parse(req.body.sauce),
-      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : { ...req.body };
-  Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Sauce modifiée !' }))
-    .catch(error => res.status(400).json({ error }));
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId === req.token.userId) {
+        const sauceObject = req.file ?
+          {
+            ...JSON.parse(req.body.sauce),
+            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+          } : { ...req.body };
+        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'Sauce modifiée !' }))
+          .catch(error => res.status(400).json({ error }));
+      }
+      else {
+        throw new Error('Unauthorized request.', { statusCode: 403 })
+      }
+    })
 };
 
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
-      const filename = sauce.imageUrl.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
-          .catch(error => res.status(400).json({ error }));
-      });
+      if (sauce.userId === req.token.userId) {
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
+            .catch(error => res.status(400).json({ error }));
+        });
+      }
+      else {
+        throw new Error('Unauthorized request.', { statusCode: 403 })
+      }
     })
     .catch(error => res.status(500).json({ error }));
 };
@@ -102,7 +115,7 @@ exports.sauceLike = (req, res, next) => {
             { $pull: { usersLiked: req.body.userId }, $inc: { likes: -1 } }
           )
             .then((sauce) => {
-              res.status(200).json({ message: "Like deleted" });
+              res.status(200).json({ message: "Like supprimé" });
             })
             .catch((error) => res.status(400).json({ error }));
         } else if (sauce.usersDisliked.includes(req.body.userId)) {
@@ -114,7 +127,7 @@ exports.sauceLike = (req, res, next) => {
             }
           )
             .then((sauce) => {
-              res.status(200).json({ message: "Dislike deleted" });
+              res.status(200).json({ message: "Dislike supprimé" });
             })
             .catch((error) => res.status(400).json({ error }));
         }
